start making post view pretty
